Rename processor to backend.
